Restructuring tests (#317)
make `/test` mimic `/src` structure-wise.

one redundant test is removed, others was just moved around, some files
are renamed.

CI are split into parallel runs.

`experimental` and some hmc tests are failing and I am going to ignore
them for now.

---------
