+++ src/model/bugsmodel.jl
@@ -505,103 +505,3 @@ function set_evaluation_mode(model::BUGSModel, mode::EvaluationMode)
     end
     return BangBang.setproperty!!(model, :evaluation_mode, mode)
 end
-
-"""
-    check_invariants(model::BUGSModel)
-
-Run a collection of sanity checks that validate the internal consistency of a
-`BUGSModel` instance.  The implementation codifies a number of *implicit
-assumptions* that the rest of the code base relies on but that are currently
-*not* part of the public interface.  Violating any of these invariants will
-very likely lead to silent logic errors further downstream, therefore it is
-recommended to call `check_invariants` after constructing or mutating a
-`BUGSModel` when refactoring. This is for testing purposes only, and ideally
-it should be removed in the future.
-
-The following conditions are verified
-
-1.  `parameters` is exactly the set of *un-observed stochastic* nodes in the
-    underlying graph `g`. (This can be violated when creating a subgraph of
-    `g`.)
-2.  The order of `parameters` matches their order of appearance in
-    `flattened_graph_node_data.sorted_nodes` (many routines assume this when
-    marching through `sorted_nodes` while consuming a flat parameter vector).
-3.  `untransformed_param_length` and `transformed_param_length` equal the sum
-    of the corresponding per-variable lengths stored in
-    `untransformed_var_lengths`/`transformed_var_lengths` for **all** current
-    parameters.
-
-The function throws an `ArgumentError` summarizing *all* detected issues (it
-does **not** abort on the first error so that users get the full picture).
-
-The function returns `true` when all checks pass so that it can be used in
-assertions inside other parts of the code base.
-"""
-
-function check_invariants(model::BUGSModel)
-    errors = String[]
-    pusherr(msg) = push!(errors, msg)
-
-    params = VarName[
-        vn for
-        vn in labels(model.g) if model.g[vn].is_stochastic && !model.g[vn].is_observed
-    ]
-    if !(Set(params) == Set(model.parameters))
-        missing_from_params = setdiff(Set(params), Set(model.parameters))
-        extra_in_params = setdiff(Set(model.parameters), Set(params))
-        if !isempty(missing_from_params)
-            pusherr("`parameters` missing nodes: $(collect(missing_from_params))")
-        end
-        if !isempty(extra_in_params)
-            pusherr(
-                "`parameters` contains nodes that are not stochastic & unobserved in the graph: $(collect(extra_in_params))",
-            )
-        end
-    end
-
-    # order consistency
-    parameter_in_order = [
-        vn for vn in model.flattened_graph_node_data.sorted_nodes if vn in model.parameters
-    ]
-    if any(parameter_in_order .!= model.parameters)
-        pusherr("Order of `parameters` does not match their order in `sorted_nodes`.")
-    end
-
-    # length bookkeeping
-    sum_untrans = sum(
-        model.untransformed_var_lengths[v] for
-        v in model.parameters if haskey(model.untransformed_var_lengths, v)
-    )
-    sum_trans = sum(
-        model.transformed_var_lengths[v] for
-        v in model.parameters if haskey(model.transformed_var_lengths, v)
-    )
-
-    if sum_untrans != model.untransformed_param_length
-        pusherr(
-            "`untransformed_param_length` (" *
-            string(model.untransformed_param_length) *
-            ") does not equal the sum of per-variable lengths (" *
-            string(sum_untrans) *
-            ").",
-        )
-    end
-    if sum_trans != model.transformed_param_length
-        pusherr(
-            "`transformed_param_length` (" *
-            string(model.transformed_param_length) *
-            ") does not equal the sum of per-variable lengths (" *
-            string(sum_trans) *
-            ").",
-        )
-    end
-
-    if !isempty(errors)
-        err_msg =
-            "The BUGSModel instance failed the following invariant checks:\n" *
-            join(errors, "\n")
-        throw(ArgumentError(err_msg))
-    end
-
-    return true
-end
