+++ test/BUGSPrimitives/distributions.jl
@@ -27,9 +27,3 @@ end
     ]
     LogDensityProblems.logdensity_and_gradient(ad_model, theta)
 end
-
-@testset "inprod(a,b)" begin
-    A = [3, 0.7, 6]
-    B = [2.6, 1.3, 3]
-    @test JuliaBUGS.BUGSPrimitives.inprod(A, B) == 26.71
-end
