+++ test/Project.toml
@@ -6,7 +6,6 @@ AdvancedMH = "5b7e9947-ddc0-4b3f-9b55-0d8042f74170"
 BangBang = "198e06fe-97b7-11e9-32a5-e1d131e6ad66"
 Bijectors = "76274a88-744f-5084-9051-94815aaf08c4"
 ChainRules = "082447d4-558c-5d27-93f4-14fc19e9eca2"
-DifferentiationInterface = "a0c0ee7d-e4b9-4e03-894e-1c5f64a51d63"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 Graphs = "86223c79-3864-5bf0-83f7-82e725a168b6"
@@ -18,7 +17,6 @@ LogDensityProblemsAD = "996a588d-648d-4e1f-a8f0-a84b347e47b1"
 MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
 MCMCChains = "c7f686f2-ff18-58e9-bc7b-31028e88f75d"
 MetaGraphsNext = "fa8bd995-216d-47f1-8a91-f3b68fbeb377"
-Mooncake = "da2b9cff-9c12-43a0-ae48-6db2b0edb7d6"
 OrderedCollections = "bac558e1-5e72-5ebc-8fee-abe8a469f55d"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 ReverseDiff = "37e2e3b7-166d-5795-8a7a-e32c996b4267"
@@ -32,7 +30,6 @@ BangBang = "0.4.1"
 AdvancedHMC = "0.6, 0.7"
 AdvancedMH = "0.8"
 ChainRules = "1"
-DifferentiationInterface = "0.6.42"
 Distributions = "0.23.8, 0.24, 0.25"
 Documenter = "0.27, 1"
 Graphs = "1"
@@ -43,7 +40,6 @@ LogDensityProblemsAD = "1"
 MacroTools = "0.5"
 MCMCChains = "6"
 MetaGraphsNext = "0.6, 0.7"
-Mooncake = "0.4"
 OrderedCollections = "1"
 ReverseDiff = "1"
 Serialization = "1.10"
