+++ test/compiler_pass.jl
@@ -173,3 +173,71 @@ end
     model = compile(model_def, data, (;))
     @test length(model.graph_evaluation_data.sorted_parameters) == 0
 end
+
+@testset "compile corner cases" begin
+    @testset "test variables exist on the left hand side of the both kinds of assignment" begin
+        @testset "not transformed variable, so error" begin
+            ex = @bugs begin
+                a ~ Normal(0, 1)
+                b = a
+                b ~ Normal(0, 1)
+            end
+            @test_throws ErrorException compile(ex, (;), (;))
+        end
+
+        @testset "transformed variable, so no error" begin
+            ex = @bugs begin
+                a ~ Normal(0, 1)
+                b = a
+                b ~ Normal(0, 1)
+            end
+            compile(ex, (; a=1), (;))
+        end
+    end
+
+    @testset "assign array variable to another array variable" begin
+        model = compile(
+            (@bugs begin
+                b[1:2] ~ dmnorm(μ[:], σ[:, :])
+                a[1:2] = b[:]
+            end), (; μ=[0, 1], σ=[1 0; 0 1]), (;)
+        )
+    end
+end
+
+@testset "error messages on undeclared variables" begin
+    @testset "same variable names refer to both scalar and array" begin
+        model_def = @bugs begin
+            x[1] ~ dnorm(0, 1)
+            y ~ dnorm(x, 1)
+        end
+        @test_throws ErrorException compile(model_def, (;))
+    end
+
+    @testset "undeclared scalar variable" begin
+        model_def = @bugs begin
+            x[1] ~ dnorm(0, 1)
+            y ~ dnorm(x[1], z)
+        end
+        @test_throws ErrorException compile(model_def, (;))
+    end
+    @testset "undeclared array variable" begin
+        model_def = @bugs begin
+            x[1] ~ dnorm(0, 1)
+            y ~ dnorm(0, x[2])
+        end
+        @test_throws ErrorException compile(model_def, (;))
+
+        model_def = @bugs begin
+            x[2] ~ dnorm(0, 1)
+            y ~ dnorm(x[1], 0)
+        end
+        @test_throws ErrorException compile(model_def, (;))
+
+        model = @bugs begin
+            x = sum(y[1:2])
+            y[1] ~ dnorm(0, 1)
+        end
+        @test_throws ErrorException compile(model, (;))
+    end
+end
