--- test/parser/test_parser.jl
@@ -1,5 +1,0 @@
-@testset "Parser" begin
-    include("bugs_macro.jl")
-    include("bugs_parser.jl")
-    include("winbugs_examples.jl")
-end
